fix important tests
